+++ dev/test_mod_impact_veneto_a.R
@@ -3,10 +3,10 @@ library(shiny)
 golem::detach_all_attached()
 golem::document_and_reload()
 
-ui <- mod_prev_pb_ui("test_1")
+ui <- mod_impact_veneto_a_ui("test_1", "A")
 
 server <- function(input,output,session){
-    mod_prev_pb_server("test_1")
+    mod_impact_veneto_a_server("test_1")
 }
 
 shinyApp(ui, server)